requirements.txt
selenium==4.10.0
webdriver_manager==3.8.6
requests==2.31.0
urllib3==2.0.3
confluent_kafka==2.1.1
kafka-python==2.0.2
python-dotenv==1.0.0
pytz==2023.3
bs4==0.0.1
kafka~=1.3.5
pyspark~=3.4.1
findspark~=2.0.1
pandas~=2.0.3
numpy~=1.25.2
py4j~=0.10.9.7
six~=1.16.0
setuptools~=65.5.1